server/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.5.0
anyio==3.7.1
apt-clone==0.2.1
apturl==0.5.2
async-timeout==4.0.3
attrs==21.2.0
beautifulsoup4==4.10.0
black==23.9.1
blinker==1.4
Brlapi==0.8.3
cachetools==5.3.1
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.0
cliapp==1.20180812.1
click==8.0.3
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
cryptography==3.4.8
cupshelpers==1.0
dataclasses-json==0.6.1
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
exceptiongroup==1.1.3
eyeD3==0.8.10
filelock==3.6.0
flake8==6.1.0
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.37
greenlet==3.0.0
grpcio==1.30.2
gyp==0.1
html5lib==1.1
httplib2==0.20.2
idna==3.3
ifaddr==0.1.7
IMDbPY==2021.4.18
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==3.2.0
keyring==23.5.0
langchain==0.0.310
langsmith==0.0.43
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==2.0.1
marshmallow==3.20.1
mccabe==0.7.0
mdurl==0.1.2
more-itertools==8.10.0
multidict==6.0.4
mypy-extensions==1.0.0
nemo-emblems==5.6.0
netaddr==0.8.0
netifaces==0.11.0
numpy==1.26.0
oauthlib==3.2.0
onboard==1.4.1
openai==0.28.1
packaging==23.2
PAM==0.4.2
pandas==2.1.1
pathspec==0.11.2
pexpect==4.8.0
Pillow==9.0.1
platformdirs==3.11.0
protobuf==4.24.4
psutil==5.9.0
ptyprocess==0.7.0
pyarrow==13.0.0
pycairo==1.20.1
pycodestyle==2.11.0
pycups==2.0.1
pycurl==7.44.1
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
pyelftools==0.27
pyflakes==3.1.0
Pygments==2.16.1
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyparted==3.11.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-gnupg==0.4.8
python-Levenshtein==0.12.2
python-magic==0.4.24
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.31.0
requests-file==1.5.1
rich==13.6.0
SecretStorage==3.3.1
setproctitle==1.2.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.3.1
SQLAlchemy==2.0.21
streamlit==1.27.2
systemd-python==234
tenacity==8.2.3
texttable==1.6.4
tinycss2==1.1.1
tldextract==3.1.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
ttystatus==0.38
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ulauncher==5.15.4
Unidecode==1.3.3
urllib3==1.26.5
validators==0.22.0
wadllib==1.3.6
watchdog==3.0.0
webencodings==0.5.1
websocket-client==1.2.3
xdg==5
xkit==0.0.0
xlrd==1.2.0
yarl==1.9.2
youtube-dl==2021.12.17
zipp==1.0.0

youtube-transcript-api
faiss-cpu
